Switch passes and update test cases
