+++ compiler/src/iree/compiler/Dialect/Flow/Transforms/DispatchLinalgOnTensors.cpp
@@ -11,9 +11,11 @@
 #include "iree-dialects/Dialect/LinalgExt/IR/LinalgExtOps.h"
 #include "iree-dialects/Dialect/LinalgExt/Passes/Transforms.h"
 #include "iree/compiler/Dialect/Flow/Conversion/TensorToFlow/Patterns.h"
+#include "iree/compiler/Dialect/Flow/Conversion/TensorToFlow/Utils.h"
 #include "iree/compiler/Dialect/Flow/IR/FlowDialect.h"
 #include "iree/compiler/Dialect/Flow/IR/FlowOps.h"
 #include "iree/compiler/Dialect/Flow/IR/FlowTypes.h"
+#include "iree/compiler/Dialect/Flow/Transforms/ConvertRegionToWorkgroups.h"
 #include "iree/compiler/Dialect/Flow/Transforms/PassDetail.h"
 #include "iree/compiler/Dialect/Flow/Transforms/Passes.h"
 #include "iree/compiler/Dialect/Flow/Transforms/RegionOpUtils.h"
@@ -63,6 +65,85 @@ static llvm::cl::opt<int> clInlineConstantByteLength(
 
 static const char kRootOpAttr[] = "__root_op__";
 static const char kFusionGroupsAttr[] = "__fused_op__";
+
+using namespace mlir;
+using namespace mlir::iree_compiler;
+using namespace mlir::iree_compiler::IREE;
+
+//===----------------------------------------------------------------------===//
+// Helpers for fusion group formation
+//===----------------------------------------------------------------------===//
+
+namespace {
+/// A rewriter that keeps track of all tensor::DimOps.
+class TensorDimTrackingRewriter : public IRRewriter {
+ public:
+  /// Create a new rewriter: Scan the given op for tensor::DimOps.
+  TensorDimTrackingRewriter(Operation *op) : IRRewriter(op->getContext()) {
+    op->walk([&](tensor::DimOp dimOp) { dimOps.insert(dimOp.getOperation()); });
+  }
+
+  /// Return all tracked tensor::DimOps.
+  SmallVector<tensor::DimOp> getTensorDimOps() {
+    SmallVector<tensor::DimOp> result;
+    for (Operation *op : dimOps) result.push_back(cast<tensor::DimOp>(op));
+    return result;
+  }
+
+ protected:
+  void notifyOperationRemoved(Operation *op) override {
+    IRRewriter::notifyOperationRemoved(op);
+    if (isa<tensor::DimOp>(op)) dimOps.erase(op);
+  }
+
+  void notifyOperationInserted(Operation *op) override {
+    IRRewriter::notifyOperationInserted(op);
+    if (isa<tensor::DimOp>(op)) dimOps.insert(op);
+  }
+
+ private:
+  SmallPtrSet<Operation *, 16> dimOps;
+};
+}  // namespace
+
+/// Simplfy the given tensor::DimOps as much as possible.
+/// * Static dimensions are replaced by constant.
+/// * Dynamic dim ops are pushed as much as possible to the top of the function,
+///   i.e., if the dim of a value is known to be equal to the dim of a value on
+///   the reverse SSA use-def chain, rewrite the value with a dim op of that
+///   value.
+static LogicalResult simplifyDimOps(RewriterBase &rewriter,
+                                    const SmallVector<tensor::DimOp> &dimOps) {
+  for (tensor::DimOp dimOp : dimOps) {
+    // Only DimOps with static indices are supported.
+    Optional<int64_t> idx = dimOp.getConstantIndex();
+    if (!idx.has_value()) continue;
+    // Only DimOps with ranked tensors are supported.
+    auto tensorType = dimOp.getSource().getType().dyn_cast<RankedTensorType>();
+    if (!tensorType) continue;
+
+    if (!tensorType.isDynamicDim(*idx)) {
+      // Rewrite static dimension with constant.
+      OpBuilder::InsertionGuard g(rewriter);
+      rewriter.setInsertionPoint(dimOp);
+      int64_t size = tensorType.getShape()[*idx];
+      rewriter.replaceOpWithNewOp<arith::ConstantIndexOp>(dimOp, size);
+      continue;
+    }
+
+    // Try to simplify dynamic dims.
+    SmallVector<Value> dynamicDims;
+    if (failed(Flow::reifyDynamicResultDims(rewriter, dimOp.getSource(),
+                                            dynamicDims)))
+      return failure();
+    unsigned ctr = 0;
+    for (int64_t i = 0; i < *dimOp.getConstantIndex(); ++i)
+      if (tensorType.isDynamicDim(i)) ++ctr;
+    rewriter.replaceOp(dimOp, dynamicDims[ctr]);
+  }
+
+  return success();
+}
 
 namespace mlir {
 namespace iree_compiler {
@@ -212,225 +293,22 @@ static SmallVector<Value> getWorkloadForRootOp(OpBuilder &builder,
   }));
 }
 
-//===----------------------------------------------------------------------===//
-// Methods that help creating the dispatch regions
-//===----------------------------------------------------------------------===//
-
-/// For an operation to be moved into the dispatch region, append `resultTypes`
-/// with the type of the results dispatch region has to return. Also
-/// append `resultDynamicDims` with values that represent the dynamic shapes of
-/// result values returned.
-static LogicalResult computeDispatchResultTypeAndDynamicDims(
-    PatternRewriter &rewriter, Operation *dispatchOp,
-    SmallVector<Type> &resultTypes, SmallVector<Value> &resultDynamicDims) {
-  auto currResultTypes = dispatchOp->getResultTypes();
-  resultTypes.append(currResultTypes.begin(), currResultTypes.end());
-  auto rankedShapedTypeOp =
-      dyn_cast<ReifyRankedShapedTypeOpInterface>(dispatchOp);
-  if (!rankedShapedTypeOp) {
-    return rewriter.notifyMatchFailure(
-        dispatchOp,
-        "expected op to implement the ReifyRankedShapedTypeOpInterface");
-  }
-
-  // Get the values for the result dims.
-  ReifiedRankedShapedTypeDims resultDims;
-  if (failed(rankedShapedTypeOp.reifyResultShapes(rewriter, resultDims))) {
-    return rewriter.notifyMatchFailure(dispatchOp,
-                                       "failed to reify shape of the result");
-  }
-  if (currResultTypes.size() != resultDims.size()) {
-    return rewriter.notifyMatchFailure(
-        dispatchOp, "expected as many result shapes as number of outputs");
-  }
-  for (auto outputType : llvm::enumerate(currResultTypes)) {
-    auto shapedOutputType = outputType.value().dyn_cast<ShapedType>();
-    if (!shapedOutputType) continue;
-    for (auto dim : llvm::enumerate(shapedOutputType.getShape())) {
-      if (ShapedType::isDynamic(dim.value())) {
-        resultDynamicDims.push_back(
-            resultDims[outputType.index()][dim.index()]);
-      }
-    }
-  }
-  return success();
-}
-
-/// Returns true if the operation has only uses in `tensor.dim` ops.
-static bool hasComputeUsesOutsideDispatch(
-    Operation *op, ArrayRef<Operation *> dispatchOps = {}) {
-  return !llvm::all_of(op->getUsers(), [&](Operation *user) {
-    return isa<tensor::DimOp>(user) || llvm::is_contained(dispatchOps, user);
-  });
-}
-
-/// Populate the workgroup count region of dispatch operations. This region
-/// mostly contains operations that are placeholders for the actual computation
-/// that will be materialized in the backends.
-static void materializeDispatchWorkgroupCountRegion(
-    OpBuilder &builder, IREE::Flow::DispatchWorkgroupsOp dispatchOp,
-    Operation *rootOp, ValueRange workload) {
-  Region &workgroupCountRegion = dispatchOp.getWorkgroupCount();
-  if (!workgroupCountRegion.empty()) return;
-
-  auto workloadArgTypes = llvm::to_vector(
-      llvm::map_range(workload, [](Value v) { return v.getType(); }));
-  Location loc = dispatchOp.getLoc();
-  SmallVector<Location> locs(workload.size(), loc);
-  OpBuilder::InsertionGuard g(builder);
-  Block *body =
-      builder.createBlock(&workgroupCountRegion, workgroupCountRegion.end(),
-                          workloadArgTypes, locs);
-  auto numWorkgroupsOp =
-      TypeSwitch<Operation *, Operation *>(rootOp)
-          .Case<IREE::LinalgExt::SetEncodingOp>([&](auto setEncodingOp) {
-            return builder
-                .create<IREE::Flow::DispatchWorkgroupCountFromSetEncodingOp>(
-                    loc, body->getArguments());
-          })
-          .Default([&](Operation *) {
-            return builder
-                .create<IREE::Flow::DispatchWorkgroupCountFromDagRootOp>(
-                    loc, body->getArguments());
-          });
-  builder.create<IREE::Flow::ReturnOp>(loc, numWorkgroupsOp->getResults());
-}
-
-/// Creates a flow.dispatch.workgroup op without arguments.
-/// All the necessary operands are transiently captured and rewritten late as
-/// operands. This greatly simplifies transformations into the resulting op.
-static FailureOr<SmallVector<Operation *>>
-buildOperandLessFlowDispatchWorkgroupOp(PatternRewriter &rewriter, Location loc,
-                                        ValueRange workload,
-                                        ArrayRef<Operation *> dispatchOps) {
-  SmallVector<Value> resultDynamicDims;
-  SmallVector<Type> resultTypes;
-
-  // 1. Compute the result types for the dispatch and the dynamic dimensions
-  //    of the result of the dispatch. If operation has only dim uses
-  //    do not make the dispatch op return those values. Those uses are
-  //    kept on the original op, and later patterns are expected to take care
-  //    of them.
-  for (auto op : dispatchOps) {
-    if (!hasComputeUsesOutsideDispatch(op, dispatchOps)) continue;
-    if (failed(computeDispatchResultTypeAndDynamicDims(
-            rewriter, op, resultTypes, resultDynamicDims))) {
-      return failure();
-    }
-  }
-
-  // 2. Create a dispatch op with just the `flow.return` terminator.
-  auto dispatchOp = rewriter.create<IREE::Flow::DispatchWorkgroupsOp>(
-      loc, workload, resultTypes, resultDynamicDims,
-      /*operands=*/ArrayRef<Value>{}, /*operandDims=*/ArrayRef<Value>{},
-      /*tiedOperands=*/ArrayRef<int64_t>{});
-  Region &region = dispatchOp.getWorkgroupBody();
-  Block *block = &region.front();
-  OpBuilder::InsertionGuard g(rewriter);
-  rewriter.setInsertionPointToEnd(block);
-  auto returnOp = rewriter.create<IREE::Flow::ReturnOp>(loc);
-  rewriter.setInsertionPoint(returnOp);
-
-  // 3. Clone the necessary operations into the dispatch and replace
-  //    all uses of the original op with the cloned op within the dispatch.
-  auto resultArgs = region.getArguments();
-  unsigned resultPos = 0;
-  unsigned resultDynamicDimsPos = 0;
-  SmallVector<Value> dispatchOpResults = dispatchOp.getResults();
-  SmallVector<Operation *> clonedOps;
-  clonedOps.reserve(dispatchOps.size());
-  for (auto op : dispatchOps) {
-    Operation *clonedOp = rewriter.clone(*op);
-    clonedOps.push_back(clonedOp);
-    rewriter.replaceOpWithinBlock(op, clonedOp->getResults(), block);
-    rewriter.setInsertionPoint(clonedOp);
-    if (!hasComputeUsesOutsideDispatch(op, dispatchOps)) continue;
-
-    // 3a. Replace all non-dim uses of the original operation with the
-    //     corresponding result of the dispatch.
-    rewriter.replaceOpWithIf(op,
-                             ArrayRef<Value>(dispatchOpResults)
-                                 .slice(resultPos, op->getNumResults()),
-                             [&](OpOperand &operand) {
-                               return !isa<tensor::DimOp>(operand.getOwner());
-                             });
-
-    // 3b. For each of the result create a `flow.dispatch.tensor.store`
-    //     operation to publish the result of the cloned operation (from within
-    //     the dispatch).
-    for (auto clonedOpResult : clonedOp->getResults()) {
-      auto resultType = clonedOpResult.getType().dyn_cast<ShapedType>();
-      if (resultType) {
-        OpBuilder::InsertionGuard g2(rewriter);
-        rewriter.setInsertionPoint(returnOp);
-        unsigned numDynamicDims = resultType.getNumDynamicDims();
-        rewriter.create<IREE::Flow::DispatchTensorStoreOp>(
-            loc, clonedOpResult, resultArgs[resultPos],
-            ArrayRef<Value>(resultDynamicDims)
-                .slice(resultDynamicDimsPos, numDynamicDims));
-        resultDynamicDimsPos += numDynamicDims;
-      }
-      resultPos++;
-    }
-  }
-
-  LLVM_DEBUG(llvm::dbgs() << "After getWorkgroupBody creation \n"
-                          << *dispatchOp << "\n");
-
-  // 4. Add a region for workgroup_count computation.
-  materializeDispatchWorkgroupCountRegion(rewriter, dispatchOp,
-                                          dispatchOps.back(), workload);
-
-  LLVM_DEBUG(llvm::dbgs() << "After workgroup_count creation \n"
-                          << *dispatchOp << "\n");
-  return clonedOps;
-}
-
-/// Returns the list of operations that are to be cloned into the dispatch
-/// based on the root operation.
-static SmallVector<Operation *> getOperationsToMoveIntoDispatch(
-    Operation *rootOp) {
-  SmallVector<Operation *> dispatchOps;
-  dispatchOps.push_back(rootOp);
-  if (!hasRootOpAttribute(rootOp)) return dispatchOps;
-
-  int64_t groupNum = getRootNumber(rootOp);
-  std::deque<Operation *> worklist;
-  worklist.push_back(rootOp);
-  llvm::SmallDenseSet<Operation *, 2> visitedOps;
-  visitedOps.insert(rootOp);
-
-  while (!worklist.empty()) {
-    Operation *currRoot = worklist.front();
-    worklist.pop_front();
-    for (auto operand : currRoot->getOperands()) {
-      auto producer = operand.getDefiningOp();
-      if (!producer || visitedOps.count(producer)) continue;
-      visitedOps.insert(producer);
-      if (!isInFusionGroup(producer, groupNum)) continue;
-      worklist.push_back(producer);
-      dispatchOps.push_back(producer);
-    }
-  }
-
-  bool sortResult = mlir::computeTopologicalSorting(dispatchOps);
-  (void)sortResult;
-  assert(sortResult && "could not compute topological sorting");
-  return llvm::to_vector(llvm::reverse(dispatchOps));
-}
-
 //===---------------------------------------------------------------------===//
 // Methods to legalize a dispatch region op, i.e. make it isolated from above.
 //===---------------------------------------------------------------------===//
 
 /// Checks if the `Value` has a use within the dispatch that is unfusable.
-static bool hasUnfusableUseInDispatch(
-    Value v, IREE::Flow::DispatchWorkgroupsOp dispatchOp) {
+static bool hasUnfusableUseInDispatch(Value v, Operation *dispatchOp) {
   for (OpOperand &use : v.getUses()) {
     Operation *user = use.getOwner();
+    Operation *ownerWorkgroupsOp =
+        user->getParentOfType<IREE::Flow::DispatchWorkgroupsOp>();
+    Operation *ownerRegionOp =
+        user->getParentOfType<IREE::Flow::DispatchRegionOp>();
+    Operation *owner = ownerWorkgroupsOp ? ownerWorkgroupsOp : ownerRegionOp;
+
     // Ignore uses outside of dispatch workgroups op.
-    if (user->getParentOfType<IREE::Flow::DispatchWorkgroupsOp>() != dispatchOp)
-      continue;
+    if (owner != dispatchOp) continue;
 
     // Cannot fuse producer of `dest` with `tensor.insert_slice`.
     if (auto insertSliceUser = dyn_cast<tensor::InsertSliceOp>(user)) {
@@ -440,300 +318,39 @@ static bool hasUnfusableUseInDispatch(
   return false;
 }
 
-/// Computes the values that will eventually be used within the dispatch
-/// workgroup op but defined outside the op after all clonable operations are
-/// cloned into the region.
-static void getUsedValuesDefinedAboveAfterCloningOps(
-    OpBuilder &builder, IREE::Flow::DispatchWorkgroupsOp dispatchOp,
-    llvm::SetVector<Value> &valuesDefinedAbove) {
-  llvm::SmallVector<Operation *> clonedOps;
+/// Collect all ops that should be cloned into the given dispatch region op.
+static SmallVector<Operation *> getCloneableOps(
+    Flow::DispatchRegionOp regionOp) {
+  // Find values that are used inside of the dispatch region but defined outside
+  // of the dispatch region.
+  llvm::SetVector<Value> valuesDefinedAbove;
+  mlir::getUsedValuesDefinedAbove(regionOp.getBody(), valuesDefinedAbove);
+  if (valuesDefinedAbove.empty()) return {};
+
+  // Traverse the defining ops of these values (and the ops on their reverse
+  // SSA use-def chain).
+  SmallVector<Operation *> result;
   llvm::SetVector<Value> visited;
   SmallVector<Value, 4> worklist;
   worklist.assign(valuesDefinedAbove.begin(), valuesDefinedAbove.end());
-  valuesDefinedAbove.clear();
   while (!worklist.empty()) {
     Value outsideValue = worklist.pop_back_val();
+    // Skip values that were already visited.
     if (visited.count(outsideValue)) continue;
     visited.insert(outsideValue);
 
     Operation *definingOp = outsideValue.getDefiningOp();
     if (!definingOp || !(isClonableIntoDispatchOp(definingOp)) ||
-        hasUnfusableUseInDispatch(outsideValue, dispatchOp)) {
+        hasUnfusableUseInDispatch(outsideValue, regionOp)) {
       valuesDefinedAbove.insert(outsideValue);
       continue;
     }
-    clonedOps.push_back(definingOp);
+    result.push_back(definingOp);
     worklist.append(definingOp->operand_begin(), definingOp->operand_end());
   }
-  // The cloned operations form a DAG. Return the cloned operations so the
-  // leaves come first, and can be cloned in-order into the dispatch region.
-  bool sortResult = mlir::computeTopologicalSorting(clonedOps);
-  (void)sortResult;
-  assert(sortResult && "could not compute topological sorting");
 
-  for (auto clonedOp : reverse(clonedOps)) {
-    Operation *clone = builder.clone(*clonedOp);
-    for (auto result : llvm::enumerate(clonedOp->getResults())) {
-      result.value().replaceUsesWithIf(
-          clone->getResult(result.index()), [&](OpOperand &use) {
-            return use.getOwner()
-                       ->getParentOfType<IREE::Flow::DispatchWorkgroupsOp>() ==
-                   dispatchOp;
-          });
-      valuesDefinedAbove.remove(result.value());
-    }
-    builder.setInsertionPoint(clone);
-  }
-
-  // Reverse the values. This is not for correctness, but more for readability
-  // of the IR.
-  llvm::SetVector<Value> reversedValues;
-  reversedValues.insert(valuesDefinedAbove.rbegin(), valuesDefinedAbove.rend());
-  std::swap(reversedValues, valuesDefinedAbove);
+  return result;
 }
-
-//===---------------------------------------------------------------------===//
-// Methods to tie operands and results of a dispatch op.
-//===---------------------------------------------------------------------===//
-
-/// Returns the tied operand for the given `resultArg`. Returns nullptr if error
-/// or not found.
-static BlockArgument getTiedOperandBlockArgument(BlockArgument resultArg) {
-  auto resultArgType =
-      resultArg.getType().dyn_cast<IREE::Flow::DispatchTensorType>();
-  if (!resultArgType ||
-      resultArgType.getAccess() != IREE::Flow::TensorAccess::WriteOnly) {
-    return nullptr;
-  }
-  // Each output block argument should just have one use.
-  if (!resultArg.hasOneUse()) return nullptr;
-
-  // And that's a flow.dispatch.output.store op.
-  auto storeOp = dyn_cast<IREE::Flow::DispatchTensorStoreOp>(
-      (*resultArg.getUses().begin()).getOwner());
-  if (!storeOp) return nullptr;
-
-  // Check if that block argument is tied to another block argument.
-  auto tieOp = storeOp.getValue().getDefiningOp<Util::TiedOpInterface>();
-  if (!tieOp) return nullptr;
-  auto tiedArg =
-      tieOp.getTiedResult(storeOp.getValue().cast<OpResult>().getResultNumber())
-          .dyn_cast_or_null<BlockArgument>();
-  if (!tiedArg) return nullptr;
-  assert(isa<IREE::Flow::DispatchWorkgroupsOp>(
-             tiedArg.getOwner()->getParentOp()) &&
-         "expected that BbArg belongs to DispatchWorkgroupsOp");
-
-  // Check that the type of the tied argument candidate and type of the output
-  // match and that the tied argument is readonly.
-  auto type = tiedArg.getType().dyn_cast<IREE::Flow::DispatchTensorType>();
-  if (!type || type.getAccess() != IREE::Flow::TensorAccess::ReadOnly ||
-      type.getBoundElementType() != resultArgType.getBoundElementType() ||
-      llvm::any_of(llvm::zip(type.getShape(), resultArgType.getShape()),
-                   [](std::tuple<int64_t, int64_t> sizes) {
-                     return std::get<0>(sizes) != ShapedType::kDynamicSize &&
-                            std::get<1>(sizes) != ShapedType::kDynamicSize &&
-                            std::get<0>(sizes) != std::get<1>(sizes);
-                   })) {
-    return nullptr;
-  }
-  return tiedArg;
-}
-
-/// Modifies `dispatchOp` to attach operand-result tie information when
-/// possible.
-static void tryToTieOperandsAndResults(
-    IREE::Flow::DispatchWorkgroupsOp dispatchOp) {
-  Block *block = dispatchOp.getBody(0);
-
-  // Go over each result to tie operand when possible, by:
-  // 1. Update the tied operand argument to take readwrite tensors.
-  // 2. Erase the result argument.
-  // 3. Attach the tie information to the DispatchWorkgroupsOp.
-  for (auto result : llvm::enumerate(dispatchOp.getResults())) {
-    if (dispatchOp.getTiedResultOperand(result.value())) continue;
-    BlockArgument outputArgument =
-        dispatchOp.getOutputBlockArgument(result.index());
-    BlockArgument tiedOperandArgument =
-        getTiedOperandBlockArgument(outputArgument);
-    if (!tiedOperandArgument) continue;
-    auto oldType =
-        tiedOperandArgument.getType().cast<IREE::Flow::DispatchTensorType>();
-    tiedOperandArgument.setType(IREE::Flow::DispatchTensorType::get(
-        IREE::Flow::TensorAccess::ReadWrite, oldType.getBoundType()));
-    outputArgument.replaceAllUsesWith(tiedOperandArgument);
-    block->eraseArgument(outputArgument.getArgNumber());
-    dispatchOp.setTiedResultOperandIndex(result.index(),
-                                         tiedOperandArgument.getArgNumber());
-  }
-}
-
-// After outlining in dispatch region we can rewrite the dispatch ops with
-// proper captures.
-static LogicalResult legalizeDispatchWorkgroupOperands(
-    IREE::Flow::DispatchWorkgroupsOp dispatchOp) {
-  Location loc = dispatchOp.getLoc();
-  Region &region = dispatchOp.getWorkgroupBody();
-  Block &block = region.front();
-  OpBuilder b = OpBuilder::atBlockBegin(&block);
-
-  llvm::SetVector<Value> valuesDefinedAbove;
-  mlir::getUsedValuesDefinedAbove(region, valuesDefinedAbove);
-  if (valuesDefinedAbove.empty()) return success();
-
-  getUsedValuesDefinedAboveAfterCloningOps(b, dispatchOp, valuesDefinedAbove);
-  b.setInsertionPointToStart(&block);
-
-  // Build a map from current operands to arguments.
-  std::pair<unsigned, unsigned> operandsIndexAndLength =
-      dispatchOp.getODSOperandIndexAndLength(1);
-  std::pair<unsigned, unsigned> operandDimsIndexAndLength =
-      dispatchOp.getODSOperandIndexAndLength(2);
-  llvm::DenseMap<Value, BlockArgument> operandToBBArg;
-  for (auto operand : llvm::enumerate(dispatchOp.getArguments())) {
-    operandToBBArg[operand.value()] = block.getArgument(operand.index());
-  }
-
-  // Of the values defined above and used in the region, add values that are not
-  // operands to the region already.
-  unsigned numOperands = operandsIndexAndLength.second;
-  unsigned numOperandDims = operandDimsIndexAndLength.second;
-  for (auto value : valuesDefinedAbove) {
-    BlockArgument bbArg = operandToBBArg.lookup(value);
-    bool wasPresent = bbArg != nullptr;
-    auto tensorType = value.getType().dyn_cast<RankedTensorType>();
-    if (!bbArg) {
-      // Create a new basic block argument for this value.
-      Type bbArgType = value.getType();
-      if (tensorType) {
-        bbArgType = IREE::Flow::DispatchTensorType::get(TensorAccess::ReadOnly,
-                                                        tensorType);
-      }
-      bbArg = block.insertArgument(numOperands, bbArgType, value.getLoc());
-    }
-
-    // Insert the operand if this is not already one.
-    if (!wasPresent) {
-      unsigned insertIdx = operandsIndexAndLength.first + numOperands;
-      dispatchOp->insertOperands(insertIdx, {value});
-      operandToBBArg[dispatchOp->getOperand(insertIdx)] = bbArg;
-      numOperands++;
-    }
-
-    Value repl = bbArg;
-    if (!wasPresent && bbArg.getType().isa<IREE::Flow::DispatchTensorType>()) {
-      // This dims for this operand does not exist. Add those.
-      SmallVector<Value> dynamicDimArgs;
-      {
-        OpBuilder::InsertionGuard g(b);
-        b.setInsertionPoint(dispatchOp);
-
-        // Fast-path for if the value comes from ops that support our dynamic
-        // shape interfaces. Otherwise we have to insert tensor.dim ops.
-        auto availableDims = IREE::Util::findDynamicDims(value);
-
-        // Add operands/args for each dynamic shape dimension.
-        SmallVector<Value> dynamicDimOperands;
-        unsigned dynamicDimIdx = 0;
-        for (auto dim : llvm::enumerate(tensorType.getShape())) {
-          if (dim.value() != ShapedType::kDynamicSize) continue;
-          if (availableDims.has_value()) {
-            dynamicDimOperands.push_back(availableDims.value()[dynamicDimIdx]);
-          } else {
-            dynamicDimOperands.push_back(b.createOrFold<tensor::DimOp>(
-                dispatchOp.getLoc(), value, dim.index()));
-          }
-          dynamicDimArgs.push_back(
-              block.insertArgument(numOperands + dynamicDimIdx,
-                                   b.getIndexType(), dispatchOp.getLoc()));
-          ++dynamicDimIdx;
-        }
-        dispatchOp->insertOperands(
-            operandsIndexAndLength.first + numOperands + numOperandDims,
-            dynamicDimOperands);
-        numOperandDims += dynamicDimOperands.size();
-        dispatchOp->insertOperands(operandsIndexAndLength.first + numOperands,
-                                   dynamicDimOperands);
-        numOperands += dynamicDimOperands.size();
-      }
-
-      // For arguments of type flow.dispatch.tensor, create a
-      // flow.dispatch.tensor.load to get the replacement values.
-      repl = b.create<IREE::Flow::DispatchTensorLoadOp>(
-          loc, value.getType().cast<RankedTensorType>(), bbArg, dynamicDimArgs);
-    }
-
-    value.replaceUsesWithIf(repl, [&](OpOperand &use) {
-      return use.getOwner()
-                 ->getParentOfType<IREE::Flow::DispatchWorkgroupsOp>() ==
-             dispatchOp;
-    });
-  }
-
-  // Update the `operand_segment_sizes`.
-  auto operandSegmentSizes = dispatchOp->getAttrOfType<DenseI32ArrayAttr>(
-      dispatchOp.getOperandSegmentSizesAttrName());
-  auto newValues = llvm::to_vector<4>(operandSegmentSizes.asArrayRef());
-  newValues[1] = numOperands;
-  newValues[2] = numOperandDims;
-  dispatchOp->setAttr(dispatchOp.getOperandSegmentSizesAttrName(),
-                      b.getDenseI32ArrayAttr(newValues));
-  return success();
-}
-
-//===----------------------------------------------------------------------===//
-// Pattern that create the dispatch region.
-//===----------------------------------------------------------------------===//
-
-namespace {
-template <typename OpType, template <typename> class Base>
-struct CreateDispatchRegionOp : Base<OpType> {
-  CreateDispatchRegionOp(MLIRContext *context,
-                         const LinalgExt::LinalgTransformationFilter &filter,
-                         PatternBenefit benefit = 1)
-      : Base<OpType>(context, benefit), transformationFilter(filter) {}
-
-  LogicalResult matchAndRewrite(OpType rootOp,
-                                PatternRewriter &rewriter) const override {
-    // TODO(ravishankarm): It is getting strange to track when to apply this
-    // pattern and when not to. Need to revisit this, with dynamic shape cases
-    // in mind.
-    if (!hasComputeUsesOutsideDispatch(rootOp)) return failure();
-    if (rootOp->template getParentOfType<IREE::Flow::DispatchWorkgroupsOp>()) {
-      return failure();
-    }
-
-    if (failed(transformationFilter.checkAndNotify(rewriter, rootOp))) {
-      return failure();
-    }
-
-    // Get the workload to use for the dispatch.
-    FailureOr<SmallVector<Value>> workload =
-        getWorkloadForRootOp(rewriter, rootOp.getOperation());
-    if (failed(workload)) {
-      return failure();
-    }
-
-    SmallVector<Operation *> dispatchOps =
-        getOperationsToMoveIntoDispatch(rootOp);
-    // Create a simple dispatch op with no operands, and not isolated from
-    // above.
-    auto clonedOps = buildOperandLessFlowDispatchWorkgroupOp(
-        rewriter, rootOp.getLoc(), workload.value(), dispatchOps);
-    if (failed(clonedOps)) {
-      return failure();
-    }
-
-    transformationFilter.replaceLinalgTransformationFilter(rewriter, rootOp);
-    transformationFilter.replaceLinalgTransformationFilter(
-        rewriter, clonedOps.value()[0]);
-    return success();
-  }
-
- private:
-  LinalgExt::LinalgTransformationFilter transformationFilter;
-};
-}  // namespace
 
 //===----------------------------------------------------------------------===//
 // Heuristics for fusing dispatchble ops with root ops using tile + fuse.
@@ -1110,6 +727,271 @@ static unsigned decideFusableLinalgOps(FunctionOpInterface funcOp,
   return numRootOps;
 }
 
+//===----------------------------------------------------------------------===//
+// Dispatch region formation
+//===----------------------------------------------------------------------===//
+
+/// Clone producers into the dispatch region.
+static LogicalResult cloneProducers(RewriterBase &rewriter,
+                                    Flow::DispatchRegionOp regionOp) {
+  SmallVector<Operation *> cloneableOps = getCloneableOps(regionOp);
+  bool sortResult = mlir::computeTopologicalSorting(cloneableOps);
+  (void)sortResult;
+  assert(sortResult && "could not compute topological sorting");
+
+  for (Operation *producer : llvm::reverse(cloneableOps))
+    if (failed(
+            clonePrecedingOpIntoDispatchRegion(rewriter, producer, regionOp)))
+      return failure();
+
+  return success();
+}
+
+static void buildSetEncodingWorkloadRegion(OpBuilder &builder, Location loc,
+                                           ArrayRef<BlockArgument> args) {
+  auto numWorkgroupsOp =
+      builder.create<Flow::DispatchWorkgroupCountFromSetEncodingOp>(loc, args);
+  builder.create<Flow::ReturnOp>(loc, numWorkgroupsOp.getResults());
+}
+
+static void buildDefaultWorkloadRegion(OpBuilder &builder, Location loc,
+                                       ArrayRef<BlockArgument> args) {
+  auto numWorkgroupsOp =
+      builder.create<Flow::DispatchWorkgroupCountFromDagRootOp>(loc, args);
+  builder.create<Flow::ReturnOp>(loc, numWorkgroupsOp.getResults());
+}
+
+/// Computes the workload and provides a workload region builder for the given
+/// root op.
+static FailureOr<Flow::WorkloadBuilder> getWorkloadBuilder(OpBuilder &builder,
+                                                           Operation *rootOp) {
+  Flow::WorkloadBuilder result;
+
+  // Compute workload (before entering the dispatch region).
+  OpBuilder::InsertionGuard g(builder);
+  SmallVector<Value> workload;
+  builder.setInsertionPoint(rootOp);
+  FailureOr<SmallVector<Value>> maybeWorkload =
+      getWorkloadForRootOp(builder, rootOp);
+  if (failed(maybeWorkload)) return failure();
+  result.workload = *maybeWorkload;
+
+  // The workload region of the WorkgroupsOp is populated by the
+  // `regionBuilder` during ConvertRegionToWorkgroups .
+  if (isa<LinalgExt::SetEncodingOp>(rootOp)) {
+    result.regionBuilder = buildSetEncodingWorkloadRegion;
+  } else {
+    result.regionBuilder = buildDefaultWorkloadRegion;
+  }
+
+  return result;
+}
+
+/// Create Flow::DispatchGroupsOps based on a fusion heuristic.
+static FailureOr<SmallVector<Flow::DispatchWorkgroupsOp>> createFusionGroups(
+    TensorDimTrackingRewriter &rewriter, FunctionOpInterface funcOp,
+    DominanceInfo const &dominanceInfo, bool generateWorkloadRegion,
+    bool aggressiveFusion) {
+  // Decide fusion groups (heuristic).
+  unsigned numRoots =
+      decideFusableLinalgOps(funcOp, dominanceInfo, aggressiveFusion);
+  SmallVector<Operation *> roots(numRoots, nullptr);
+  DenseMap<unsigned, SmallVector<Operation *>> producers;
+
+  // TODO: Incrementally add ops to an empty DispatchGroupOp instead of
+  // annotating fusion group IDs via attributes.
+  funcOp.walk([&](Operation *op) {
+    if (hasRootOpAttribute(op)) {
+      roots[getRootNumber(op)] = op;
+      removeRootOpAttribute(op);
+    }
+    if (hasFusionGroupsAttribute(op)) {
+      assert(getFusionGroups(op).size() == 1 && "expected exactly one group");
+      producers[getFusionGroups(op).front()].push_back(op);
+      removeFusionGroupsAttribute(op);
+    }
+  });
+
+  // Create a DispatchRegionOp for every fusion group.
+  OpBuilder::InsertionGuard g(rewriter);
+  SmallVector<Flow::DispatchRegionOp> regionOps;
+  DenseMap<Flow::DispatchRegionOp, Optional<Flow::WorkloadBuilder>>
+      workloadBuilders;
+  for (const auto &it : llvm::enumerate(roots)) {
+    // Compute workload.
+    Optional<Flow::WorkloadBuilder> workloadBuilder = llvm::None;
+    if (generateWorkloadRegion) {
+      auto maybeBuilder = getWorkloadBuilder(rewriter, /*rootOp=*/it.value());
+      if (failed(maybeBuilder)) return failure();
+      workloadBuilder = *maybeBuilder;
+    }
+
+    // Simplify tensor::DimOps.
+    SmallVector<tensor::DimOp> dimOps = rewriter.getTensorDimOps();
+    if (failed(simplifyDimOps(rewriter, dimOps))) return failure();
+
+    // Create fusion group.
+    Flow::DispatchRegionOp regionOp;
+    auto maybeRegionOp = Flow::wrapOpInDispatchRegion(rewriter, it.value());
+    if (failed(maybeRegionOp)) return failure();
+    regionOp = *maybeRegionOp;
+
+    // Sort producers topologically. All producers must be in the same block as
+    // the root.
+    bool sortResult = mlir::computeTopologicalSorting(producers[it.index()]);
+    (void)sortResult;
+    assert(sortResult && "could not compute topological sorting");
+
+    // Move ops into the region.
+    for (Operation *producer : llvm::reverse(producers[it.index()])) {
+      auto newRegionOp =
+          movePrecedingOpIntoDispatchRegion(rewriter, producer, regionOp);
+      if (failed(newRegionOp)) return failure();
+      regionOp = *newRegionOp;
+    }
+
+    workloadBuilders[regionOp] = workloadBuilder;
+    regionOps.push_back(regionOp);
+  }
+
+  // Clone additional producers and rewrite to DispatchWorkgroupsOp.
+  SmallVector<Flow::DispatchWorkgroupsOp> result;
+  for (auto regionOp : regionOps) {
+    if (failed(cloneProducers(rewriter, regionOp))) return failure();
+    auto maybeWorkgroupOp =
+        Flow::rewriteFlowDispatchRegionToFlowDispatchWorkgroups(
+            regionOp, rewriter, workloadBuilders[regionOp]);
+    if (failed(maybeWorkgroupOp)) return failure();
+
+    result.push_back(*maybeWorkgroupOp);
+  }
+
+  return result;
+}
+
+/// Wrap a single op in a DispatchWorkgroupsOp.
+static FailureOr<Flow::DispatchWorkgroupsOp> wrapInWorkgroupsOp(
+    TensorDimTrackingRewriter &rewriter, Operation *op,
+    bool generateWorkloadRegion) {
+  // Compute workload.
+  Optional<Flow::WorkloadBuilder> workloadBuilder = llvm::None;
+  if (generateWorkloadRegion) {
+    auto maybeBuilder = getWorkloadBuilder(rewriter, op);
+    if (failed(maybeBuilder)) return failure();
+    workloadBuilder = *maybeBuilder;
+  }
+
+  // Simplify tensor::DimOps.
+  SmallVector<tensor::DimOp> dimOps = rewriter.getTensorDimOps();
+  if (failed(simplifyDimOps(rewriter, rewriter.getTensorDimOps())))
+    return failure();
+
+  // Wrap operation.
+  auto regionOp = Flow::wrapOpInDispatchRegion(rewriter, op);
+  if (failed(regionOp)) return failure();
+  if (failed(cloneProducers(rewriter, *regionOp))) return failure();
+  auto workgroupsOp = Flow::rewriteFlowDispatchRegionToFlowDispatchWorkgroups(
+      *regionOp, rewriter, workloadBuilder);
+  if (failed(workgroupsOp)) return failure();
+  return *workgroupsOp;
+}
+
+/// Wrap all given ops in a DispatchWorkgroupsOp.
+static FailureOr<SmallVector<Flow::DispatchWorkgroupsOp>> wrapInWorkgroupsOp(
+    TensorDimTrackingRewriter &rewriter, SmallVector<Operation *> rootOps,
+    bool generateWorkloadRegion) {
+  SmallVector<Flow::DispatchWorkgroupsOp> result;
+  for (Operation *rootOp : rootOps) {
+    auto workgroupsOp =
+        wrapInWorkgroupsOp(rewriter, rootOp, generateWorkloadRegion);
+    if (failed(workgroupsOp)) return failure();
+    result.push_back(*workgroupsOp);
+  }
+  return result;
+}
+
+/// Wrap all ops of the given types that are direct children of the given op in
+/// DispatchWorkgroupsOps.
+template <typename... OpTys>
+static FailureOr<SmallVector<Flow::DispatchWorkgroupsOp>> wrapInWorkgroupsOp(
+    TensorDimTrackingRewriter &rewriter, Operation *op,
+    bool generateWorkloadRegion) {
+  // Find ops of type OpTys.
+  SmallVector<Operation *> rootOps;
+  for (Region &r : op->getRegions())
+    for (Block &b : r.getBlocks())
+      for (Operation &op : b)
+        if (isa<OpTys...>(&op)) rootOps.push_back(&op);
+
+  // Wrap ops in DispatchWorkgroupsOps.
+  return wrapInWorkgroupsOp(rewriter, rootOps, generateWorkloadRegion);
+}
+
+/// Return `true` if the given op is contained in DispatchWorkgroupsOp or in a
+/// DispatchRegionOp.
+static bool isInDispatchRegion(Operation *op) {
+  return op->getParentOfType<Flow::DispatchWorkgroupsOp>() ||
+         op->getParentOfType<Flow::DispatchRegionOp>();
+}
+
+/// Rewrite top-level InsertSliceOps to FlowUpdateOps or wrap them in a
+/// dispatch region.
+LogicalResult convertInsertSliceOps(
+    TensorDimTrackingRewriter &rewriter, mlir::FunctionOpInterface funcOp,
+    SmallVector<Flow::DispatchWorkgroupsOp> &workgroupsOps,
+    bool generateWorkloadRegion) {
+  // Find eligible InsertSliceOps.
+  SmallVector<tensor::InsertSliceOp> insertSliceOps;
+  funcOp.walk([&](tensor::InsertSliceOp op) {
+    if (!isInDispatchRegion(op)) insertSliceOps.push_back(op);
+  });
+
+  // Rewrite InsertSliceOps to FlowUpdateOps.
+  SmallVector<Operation *> remainingInsertSliceOps;
+  for (tensor::InsertSliceOp insertSliceOp : insertSliceOps)
+    if (failed(
+            Flow::convertInsertSliceOpToFlowUpdateOp(rewriter, insertSliceOp)))
+      remainingInsertSliceOps.push_back(insertSliceOp);
+
+  // Create a DispatchWorkgroupsOp for every remaining InsertSliceOp.
+  FailureOr<SmallVector<Flow::DispatchWorkgroupsOp>> newWorkgroupsOps =
+      wrapInWorkgroupsOp(rewriter, remainingInsertSliceOps,
+                         generateWorkloadRegion);
+  if (failed(newWorkgroupsOps)) return failure();
+  workgroupsOps.append(newWorkgroupsOps->begin(), newWorkgroupsOps->end());
+
+  return success();
+}
+
+/// Rewrite top-level ExtractSliceOps to FlowSliceOps or wrap them in a
+/// dispatch region.
+LogicalResult convertExtractSliceOps(
+    TensorDimTrackingRewriter &rewriter, mlir::FunctionOpInterface funcOp,
+    SmallVector<Flow::DispatchWorkgroupsOp> &workgroupsOps,
+    bool generateWorkloadRegion) {
+  // Find eligible ExtractSliceOps.
+  SmallVector<tensor::ExtractSliceOp> extractSliceOps;
+  funcOp.walk([&](tensor::ExtractSliceOp op) {
+    if (!isInDispatchRegion(op)) extractSliceOps.push_back(op);
+  });
+
+  // Rewrite ExtractSliceOps to FlowSliceOps.
+  SmallVector<Operation *> remainingExtractSliceOps;
+  for (tensor::ExtractSliceOp extractSliceOp : extractSliceOps)
+    if (failed(
+            Flow::convertExtractSliceOpToFlowSliceOp(rewriter, extractSliceOp)))
+      remainingExtractSliceOps.push_back(extractSliceOp);
+
+  // Create a DispatchWorkgroupsOp for every remaining ExtractSliceOp.
+  FailureOr<SmallVector<Flow::DispatchWorkgroupsOp>> newWorkgroupsOps =
+      wrapInWorkgroupsOp(rewriter, remainingExtractSliceOps,
+                         generateWorkloadRegion);
+  if (failed(newWorkgroupsOps)) return failure();
+  workgroupsOps.append(newWorkgroupsOps->begin(), newWorkgroupsOps->end());
+
+  return success();
+}
+
 namespace {
 /// Pass declaration.
 struct DispatchLinalgOnTensorsPass
@@ -1119,11 +1001,14 @@ struct DispatchLinalgOnTensorsPass
         .insert<AffineDialect, IREE::Flow::FlowDialect, linalg::LinalgDialect,
                 scf::SCFDialect, tensor::TensorDialect>();
   }
-  DispatchLinalgOnTensorsPass(bool aggressiveFusion) {
+  DispatchLinalgOnTensorsPass(bool aggressiveFusion,
+                              bool generateWorkloadRegion) {
     this->aggressiveFusion = aggressiveFusion;
+    this->generateWorkloadRegion = generateWorkloadRegion;
   }
   DispatchLinalgOnTensorsPass(const DispatchLinalgOnTensorsPass &pass)
-      : DispatchLinalgOnTensorsPass(pass.aggressiveFusion) {}
+      : DispatchLinalgOnTensorsPass(pass.aggressiveFusion,
+                                    pass.generateWorkloadRegion) {}
   void runOnOperation() override;
 
  private:
@@ -1132,106 +1017,19 @@ struct DispatchLinalgOnTensorsPass
 };
 }  // namespace
 
-/// For all ops within `funcOp` tagged as root ops, create dispatch regions.
-LogicalResult createDispatchRegionsFromRootOps(mlir::Operation *funcOp,
-                                               RewritePatternSet &&patterns) {
-  MLIRContext *context = funcOp->getContext();
-
-  // Create the dispatch region, first without the isolate region from above
-  // property.
-  if (failed(applyPatternsAndFoldGreedily(funcOp, std::move(patterns)))) {
-    return failure();
-  }
-
-  // Run canonicalization patterns and pattern to resolve tensor.dim of result
-  // values into tensor.dim of its operands..
-  RewritePatternSet canonicalizationPatterns(context);
-  memref::populateResolveRankedShapeTypeResultDimsPatterns(
-      canonicalizationPatterns);
-  if (failed(applyPatternsAndFoldGreedily(
-          funcOp, std::move(canonicalizationPatterns)))) {
-    return failure();
-  }
-
-  LLVM_DEBUG({
-    llvm::dbgs() << "\n--- After dispatch op formation ---\n";
-    funcOp->print(llvm::dbgs(), OpPrintingFlags().useLocalScope());
-    llvm::dbgs() << "\n\n";
-  });
-
-  // After outlining in dispatch region we can rewrite the dispatch ops with
-  // proper captures to make it isolated from above.
-  if (funcOp
-          ->walk([&](IREE::Flow::DispatchWorkgroupsOp op) -> WalkResult {
-            return legalizeDispatchWorkgroupOperands(op);
-          })
-          .wasInterrupted()) {
-    return failure();
-  }
-
-  LLVM_DEBUG({
-    llvm::dbgs() << "\n--- After dispatch op legalization ---\n";
-    funcOp->print(llvm::dbgs(), OpPrintingFlags().useLocalScope());
-    llvm::dbgs() << "\n\n";
-  });
-
-  // Now try to see if we can tie certain results to operands in order to
-  // indicate sharing storage. This need to happen here because it needs to
-  // access region block arguments for input/output tensors, which aren't
-  // available until now.
-  funcOp->walk([&](IREE::Flow::DispatchWorkgroupsOp op) {
-    tryToTieOperandsAndResults(op);
-  });
-
-  LLVM_DEBUG({
-    llvm::dbgs() << "\n--- After tieing operands and results ---\n";
-    funcOp->print(llvm::dbgs(), OpPrintingFlags().useLocalScope());
-    llvm::dbgs() << "\n\n";
-  });
-
-  // Finally fold `tensor.insert_slice/extract_slice` operations with
-  // `flow.dispatch.tensor.load/store`.
-  RewritePatternSet foldExtractInsertSliceOps(context);
-  populateTensorSliceOpWithDispatchTensorOpFoldingPatterns(
-      foldExtractInsertSliceOps, context);
-  if (failed(applyPatternsAndFoldGreedily(
-          funcOp, std::move(foldExtractInsertSliceOps)))) {
-    return failure();
-  }
-
-  return success();
-}
-
 void DispatchLinalgOnTensorsPass::runOnOperation() {
-  auto funcOp = getOperation();
+  mlir::FunctionOpInterface funcOp = getOperation();
   MLIRContext *context = &getContext();
+
   DominanceInfo const &dominanceInfo = getAnalysis<DominanceInfo>();
-  decideFusableLinalgOps(funcOp, dominanceInfo, aggressiveFusion);
+  TensorDimTrackingRewriter rewriter(funcOp);
 
-  LLVM_DEBUG({
-    llvm::dbgs() << "\n--- After annotating linalg op fusion scheme ---\n";
-    funcOp->print(llvm::dbgs(), OpPrintingFlags().useLocalScope());
-    llvm::dbgs() << "\n\n";
-  });
-
-  {
-    LinalgExt::LinalgTransformationFilter filterForComputeOps(
-        [](Operation *op) { return success(hasRootOpAttribute(op)); }, {},
-        StringAttr::get(context, "indispatch"));
-    filterForComputeOps.setMatchByDefault();
-    RewritePatternSet computeOpDispatchPatterns(context);
-    computeOpDispatchPatterns.insert<
-        CreateDispatchRegionOp<TilingInterface, OpInterfaceRewritePattern>,
-        CreateDispatchRegionOp<LinalgExt::SetEncodingOp, OpRewritePattern>,
-        CreateDispatchRegionOp<LinalgExt::UnsetEncodingOp, OpRewritePattern>,
-        CreateDispatchRegionOp<tensor::InsertSliceOp, OpRewritePattern>>(
-        context, filterForComputeOps);
-    if (failed(createDispatchRegionsFromRootOps(
-            funcOp, std::move(computeOpDispatchPatterns)))) {
-      funcOp.emitOpError("failed to create dispatch region based on operation");
-      return signalPassFailure();
-    }
-  }
+  // Step 1: Create a DispatchWorkgroupsOp for every fusion group.
+  auto maybeWorkgroupsOps =
+      createFusionGroups(rewriter, funcOp, dominanceInfo,
+                         generateWorkloadRegion, aggressiveFusion);
+  if (failed(maybeWorkgroupsOps)) return signalPassFailure();
+  SmallVector<Flow::DispatchWorkgroupsOp> workgroupsOps = *maybeWorkgroupsOps;
 
   LLVM_DEBUG({
     llvm::dbgs() << "\n--- After first step of dispatch region formation ---\n";
@@ -1239,69 +1037,52 @@ void DispatchLinalgOnTensorsPass::runOnOperation() {
     llvm::dbgs() << "\n\n";
   });
 
-  /// Convert remaining ops to Flow ops.
+  // Step 2: Rewrite InsertSliceOps to FlowUpdateOps.
+  if (failed(convertInsertSliceOps(rewriter, funcOp, workgroupsOps,
+                                   generateWorkloadRegion)))
+    return signalPassFailure();
+
+  // Step 3: Rewrite ExtractSliceOps to FlowUpdateOps.
+  if (failed(convertExtractSliceOps(rewriter, funcOp, workgroupsOps,
+                                    generateWorkloadRegion)))
+    return signalPassFailure();
+
+  // Step 4: Create a DispatchWorkgroupsOp for certain other ops.
+  FailureOr<SmallVector<Flow::DispatchWorkgroupsOp>> newWorkgroupsOps =
+      wrapInWorkgroupsOp<LinalgExt::SetEncodingOp, LinalgExt::UnsetEncodingOp>(
+          rewriter, funcOp, generateWorkloadRegion);
+  if (failed(newWorkgroupsOps)) return signalPassFailure();
+  workgroupsOps.append(newWorkgroupsOps->begin(), newWorkgroupsOps->end());
+
+  // A few extra canonicalizations/lowerings.
   {
     RewritePatternSet convertToFlowPatterns(context);
-    populateTensorToFlowConversionPatterns(context, convertToFlowPatterns);
+    Flow::populateTensorToFlowConversionPatterns(context,
+                                                 convertToFlowPatterns);
     memref::populateResolveRankedShapeTypeResultDimsPatterns(
         convertToFlowPatterns);
     IREE::Flow::TensorReshapeOp::getCanonicalizationPatterns(
         convertToFlowPatterns, context);
+    if (failed(applyPatternsAndFoldGreedily(funcOp,
+                                            std::move(convertToFlowPatterns))))
+      return signalPassFailure();
+
+    // Finally fold `tensor.insert_slice/extract_slice` operations with
+    // `flow.dispatch.tensor.load/store`.
+    RewritePatternSet foldExtractInsertSliceOps(context);
+    Flow::populateTensorSliceOpWithDispatchTensorOpFoldingPatterns(
+        foldExtractInsertSliceOps, context);
     if (failed(applyPatternsAndFoldGreedily(
-            funcOp, std::move(convertToFlowPatterns)))) {
-      funcOp.emitOpError(
-          "failed to convert operation outside dispatch region into flow ops");
+            funcOp, std::move(foldExtractInsertSliceOps))))
       return signalPassFailure();
-    }
   }
-
-  /// Move yet more remaining ops into dispatch region.
-
-  // Start with just moving the tensor.insert_slice into its dispatch.
-  {
-    LinalgExt::LinalgTransformationFilter filterForInsertSliceOps(
-        ArrayRef<StringAttr>{}, StringAttr::get(context, "indispatch"));
-    RewritePatternSet insertSliceOpDispatchPatterns(context);
-    insertSliceOpDispatchPatterns.insert<
-        CreateDispatchRegionOp<tensor::InsertSliceOp, OpRewritePattern>>(
-        context, filterForInsertSliceOps);
-    if (failed(createDispatchRegionsFromRootOps(
-            funcOp, std::move(insertSliceOpDispatchPatterns)))) {
-      funcOp.emitOpError(
-          "failed to move non-contiguous insert_slice ops into dispatch "
-          "region");
-      return signalPassFailure();
-    }
-  }
-
-  // Now move all remaining ops that need to be cleaned up.
-  {
-    LinalgExt::LinalgTransformationFilter filterForCleanupOps(
-        ArrayRef<StringAttr>{}, StringAttr::get(context, "indispatch"));
-    RewritePatternSet cleanUpDispatchPatterns(context);
-    cleanUpDispatchPatterns.insert<
-        CreateDispatchRegionOp<tensor::ExtractSliceOp, OpRewritePattern>>(
-        context, filterForCleanupOps);
-    if (failed(createDispatchRegionsFromRootOps(
-            funcOp, std::move(cleanUpDispatchPatterns)))) {
-      funcOp.emitOpError(
-          "failed to move non-contiguous extract_slice ops into dispatch "
-          "region");
-      return signalPassFailure();
-    }
-  }
-
-  // Finally walk all the ops and remove the attributes
-  funcOp.walk([](Operation *op) {
-    removeFusionGroupsAttribute(op);
-    removeRootOpAttribute(op);
-    op->removeAttr(IREE::LinalgExt::LinalgTransforms::kLinalgTransformMarker);
-  });
 }
 
 std::unique_ptr<InterfacePass<mlir::FunctionOpInterface>>
-createDispatchLinalgOnTensorsPass(bool aggressiveFusion) {
-  return std::make_unique<DispatchLinalgOnTensorsPass>(aggressiveFusion);
+createDispatchLinalgOnTensorsPass(bool aggressiveFusion,
+                                  bool generateWorkloadRegion) {
+  return std::make_unique<DispatchLinalgOnTensorsPass>(aggressiveFusion,
+                                                       generateWorkloadRegion);
 }
 
 }  // namespace Flow
