+++ compiler/src/iree/compiler/Dialect/Flow/Transforms/Passes.td
@@ -75,13 +75,9 @@ def DispatchLinalgOnTensors :
   let options = [
     Option<"aggressiveFusion", "aggressive-fusion", "bool",
            /*default=*/"false", "Fuse with aggressive heuristics">,
+    Option<"generateWorkloadRegion", "genereate-workload-region", "bool",
+           /*default=*/"true", "Generate workload regions of WorkgroupOps">,
   ];
-}
-
-def DispatchLinalgOnTensorsViaRegionOps :
-    InterfacePass<"iree-flow-dispatch-linalg-on-tensors-via-regionops-pass", "mlir::FunctionOpInterface"> {
-  let summary = "Dispatch Linalg operations on tensors by using tile and distribute (via DispatchRegionOps)";
-  let constructor = "mlir::iree_compiler::IREE::Flow::createDispatchLinalgOnTensorsViaRegionOpsPass()";
 }
 
 def DispatchWithTransformDialect :
