+++ tests/transform_dialect/cpu/matmul.mlir
@@ -13,8 +13,7 @@ func.func @matmul_static(
 // Run with C++ dispatch region formation but transform dialect codegen
 // RUN: iree-opt %s --iree-hal-target-backends=llvm-cpu \
 // RUN:   --iree-abi-transformation-pipeline --iree-flow-transformation-pipeline \
-// RUN:   --iree-flow-dispatch-via-region-ops \
-// RUN:   --iree-flow-dispatch-via-region-ops-generate-workload-region=false \
+// RUN:   --iree-flow-dispatch-generate-workload-region=false \
 // RUN:   --iree-stream-transformation-pipeline \
 // RUN:   --iree-hal-configuration-pipeline | \
 // RUN: iree-opt --pass-pipeline='builtin.module(hal.executable(hal.executable.variant(iree-llvmcpu-lower-executable-target)))' \
